updating branch

+++ src/main/java/com/application/repositories/EnseignantRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.application.models.Enseignant;
 
 @Repository
-public interface EnseignantRepository extends JpaRepository<Enseignant,Integer> {
+public interface EnseignantRepository extends JpaRepository<Enseignant,Long> {
 
 }
